cart checkout

+++ public/front/js/cart.js
@@ -3,28 +3,33 @@
         var _this = $(this);
         var id = $(this).attr('product-id');
         var _url = info['destroyUrl'] + '/' + id;
-        $.ajax({
-            type: 'DELETE',
-            url: _url,
-            data: {id: id}
-        })
-        .done(function (data) {
-            _this.parent().parent().remove();
-            $('span.cart-amunt, #sub-total, #total').text(data['totalPrice']);
-            $('span.product-count').text(data['totalItems']);
+        swal({
+            title: '',
+            text: info['remove'],
+            type: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#DD6B55",
+            confirmButtonText: info['yes'],
+            cancelButtonText: info['no']
+        }, function(isConfirm){
+            if (isConfirm) {
+                $.ajax({
+                    type: 'DELETE',
+                    url: _url,
+                    data: {id: id}
+                })
+                .done(function (data) {
+                    _this.parent().parent().remove();
+                    $('span.cart-amunt, #sub-total, #total, .nav-cart span').text(data['totalPrice']);
+                    $('span.product-count').text(data['totalItems']);
+                })
+            }
         })
     });
     $(".btn-number").on('click', function () {
         $("input#update-cart").attr('disabled', false);
         var price = $(this).parent().parent().find('.product-price').find('.amount').attr('data-price');
         var quantity = $(this).parent().find('.input-number').val();
-
-        if ($(this).attr('data-type') == 'minus') {
-            quantity -= 1;
-        } else if ($(this).attr('data-type') == 'plus') {
-            quantity = Number(quantity) + 1;
-        }
-
         var total = (quantity * price * info['currency_unit']).toLocaleString() + info['currency'];
        $(this).parent().parent().find('.product-subtotal').find('.amount').text(total);
     });
@@ -43,7 +48,8 @@
             data: data,
         })
         .done(function (data) {
-            $('span.cart-amunt, #sub-total, #total').text(data['totalPrice']);
+            $('span.cart-amunt, #sub-total, #total, .nav-cart span').text(data['totalPrice']);
             $('span.product-count').text(data['totalItems']);
+            swal('', info['updated'], 'success');
         });
     });
